Create tests for product-service

+++ product-service/src/main/java/ru/alex9043/productservice/service/impl/ProductServiceImpl.java
@@ -61,6 +61,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteProduct(UUID id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Продукта с id: " + id + " не существует");
+        }
         repository.deleteById(id);
     }
 }
